Create Art Store Database and Stats
This code was created to show an inventory database for a mock art supply store. It assigns an id to each item alone with the quantity, price, and brand.
The final table is set to show all the items with just their price and brand showing while in descending order by price.